Revert "close #1968 implement CRUD template_guests with softdelete (#1981)"
This reverts commit 8a2f51bd88e0495d5f263a7da2d884341c591284.
