+++ app/controllers/spree/api/v2/storefront/id_cards_controller.rb
@@ -4,17 +4,27 @@ module Spree
       module Storefront
         class IdCardsController < ::Spree::Api::V2::ResourceController
           def create
-            context = id_card_manager_context
+            context = SpreeCmCommissioner::GuestIdCardManager.call(
+              card_type: id_card_params[:card_type],
+              front_image_url: id_card_params[:front_image_url],
+              back_image_url: id_card_params[:back_image_url],
+              guest_id: id_card_params[:guest_id]
+            )
 
-            if context&.success?
+            if context.success?
               render_serialized_payload { serialize_resource(context.result) }
             else
-              render_error_payload(context&.message || 'Failed to create ID card')
+              render_error_payload(context.message)
             end
           end
 
           def update
-            context = id_card_manager_context
+            context = SpreeCmCommissioner::GuestIdCardManager.call(
+              card_type: id_card_params[:card_type],
+              front_image_url: id_card_params[:front_image_url],
+              back_image_url: id_card_params[:back_image_url],
+              guest_id: id_card_params[:guest_id]
+            )
 
             if context.success?
               render_serialized_payload { serialize_resource(context.result) }
@@ -36,22 +46,7 @@ module Spree
           private
 
           def id_card_params
-            params.permit(:card_type, :front_image_url, :back_image_url, :guest_id, :template_guest_id)
-          end
-
-          def id_card_manager_context
-            options = {
-              card_type: id_card_params[:card_type],
-              front_image_url: id_card_params[:front_image_url],
-              back_image_url: id_card_params[:back_image_url]
-            }
-            if id_card_params[:guest_id].present?
-              id_card_options = options.merge(guestable_id: id_card_params[:guest_id])
-              SpreeCmCommissioner::GuestIdCardManager.call(**id_card_options)
-            elsif id_card_params[:template_guest_id].present?
-              id_card_options = options.merge(guestable_id: id_card_params[:template_guest_id])
-              SpreeCmCommissioner::TemplateGuestIdCardManager.call(**id_card_options)
-            end
+            params.permit(:card_type, :front_image_url, :back_image_url, :guest_id)
           end
 
           def model_class
