--- app/interactors/spree_cm_commissioner/template_guest_id_card_manager.rb
@@ -1,7 +1,0 @@
-module SpreeCmCommissioner
-  class TemplateGuestIdCardManager < IdCardManager
-    def guestable
-      @guestable ||= SpreeCmCommissioner::TemplateGuest.find(context.guestable_id)
-    end
-  end
-end
