+++ app/models/spree_cm_commissioner/id_card.rb
@@ -1,11 +1,8 @@
 module SpreeCmCommissioner
   class IdCard < SpreeCmCommissioner::Base
-    # Polymorphic association
-    belongs_to :id_cardable, polymorphic: true
-
     enum card_type: { :national_id_card => 0, :passport => 1, :student_id_card => 2 }
 
-    # belongs_to :guest, class_name: 'SpreeCmCommissioner::Guest'
+    belongs_to :guest, class_name: 'SpreeCmCommissioner::Guest'
 
     has_one :front_image, as: :viewable, dependent: :destroy, class_name: 'SpreeCmCommissioner::FrontImage'
     has_one :back_image, as: :viewable, dependent: :destroy, class_name: 'SpreeCmCommissioner::BackImage'
