+++ config/routes.rb
@@ -458,9 +458,6 @@ Spree::Core::Engine.add_routes do
         resources :guests, only: %i[create update show] do
           resources :id_cards
         end
-        resources :template_guests, only: %i[index create show update destroy] do
-          resources :id_cards
-        end
         resources :user_guests do
           resources :id_cards
         end
