--- db/migrate/20240926120947_add_template_guest_id_to_spree_cm_commissioner_guest.rb
@@ -1,5 +1,0 @@
-class AddTemplateGuestIdToSpreeCmCommissionerGuest < ActiveRecord::Migration[7.0]
-  def change
-    add_reference :cm_guests, :template_guest, foreign_key: { to_table: :cm_template_guests }, index: true, if_not_exists: true
-  end
-end
