--- db/migrate/20240927082749_migrate_guest_id_to_polymorphic_in_cm_id_cards.rb
@@ -1,9 +1,0 @@
-class MigrateGuestIdToPolymorphicInCmIdCards < ActiveRecord::Migration[7.0]
-  def up
-    add_reference :cm_id_cards, :id_cardable, polymorphic: true, index: true
-  end
-
-  def down
-    remove_reference :cm_id_cards, :id_cardable, polymorphic: true, index: true
-  end
-end
